Add getters and constructors

+++ Lab-2/src/Model/Vegetable.java
@@ -1,6 +1,6 @@
 package Model;
 
 public interface Vegetable {
-    int get_weight();
-    int get_price();
+    double get_weight();
+    double get_price();
 }
